refactor: extract shared and demo utilities
