+++ src/core/utils.js
@@ -1,0 +1,15 @@
+export function getMousePos(e) {
+  const x = e.changedTouches ? e.changedTouches[0].clientX : e.clientX;
+  const y = e.changedTouches ? e.changedTouches[0].clientY : e.clientY;
+  const target = e.target;
+
+  return {
+    x,
+    y,
+    target,
+  };
+}
+
+export function mix(a, b, t) {
+  return a * (1 - t) + b * t;
+}
